src/medipal/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.3.1
pypdf==3.17.4
python-docx==1.1.0
easyocr==1.7.0
Pillow==10.2.0
python-multipart==0.0.9
transformers==4.38.2
llama-cpp-python==0.2.55     # CPU quantized
tavily-python==0.3.1
python-dotenv==0.19.2
pydantic==2.6.0
google-cloud-vision==3.6.0
google-auth==2.23.0
google-api-core==2.11.0